Replace drawer by modal
