+++ public/app/features/datasources/components/picker/utils.ts
@@ -1,0 +1,17 @@
+import { DataSourceInstanceSettings, DataSourceRef } from '@grafana/data';
+
+export function isDataSourceMatch(
+  ds: DataSourceInstanceSettings | undefined,
+  current: string | DataSourceInstanceSettings | DataSourceRef | null | undefined
+): boolean | undefined {
+  if (!ds) {
+    return false;
+  }
+  if (!current) {
+    return false;
+  }
+  if (typeof current === 'string') {
+    return ds.uid === current;
+  }
+  return ds.uid === current.uid;
+}
